Allow floating point recoil values and fix config parsing

+++ src/recoil/Structs.h
@@ -2,8 +2,8 @@
 
 struct WeaponRecoil
 {
-    int Vertical;
-    int Horizontal;
+    float Vertical;
+    float Horizontal;
 };
 
 enum class ScopeType
